<feat> Add user data to MainCalendar and send date data to server
This commit adds a new interface for User, which is then passed as a prop to the MainCalendar component in MainPage. The addData function in MainCalendar now sends date information along with the user ID to the server using a POST request. The server-side controller has been updated accordingly to handle this new data.