<feat> Add admin shift dates to calendar
This commit adds functionality to allow admins to add shifts to the calendar by clicking on a specific shift. The AdminOneShift component now includes a function that updates the state of adminDates with the new shift data. The AdminShiftsList component now passes down setAdminDates and adminDates as props, which are used in AdminOneShift. Finally, the events prop is added to FullCalendar in AdminPanel, which displays all shifts added by admins onto the calendar.