<feat> Add sendToDatabase function to AdminNav component
This commit adds a new function called sendToDatabase to the AdminNav component. This function sends data to the server using fetch API and saves it in the database. The data is passed as props from the parent component. Additionally, this commit modifies the readyWorkShifts controller on the server-side to accept an array of objects instead of individual values.